requirements.txt
django==1.11.3
PyMySQL==0.8.0
Pillow